Rename countyCodes -> counties
[#146178793]
