+++ app/javascript/common/CountySelectField.jsx
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types'
 class CountySelectField extends React.Component {
   render() {
     const {
-      countyCodes,
+      counties,
       gridClassName,
       id,
       onChange,
@@ -21,14 +21,14 @@ class CountySelectField extends React.Component {
         value={value}
       >
         <option key='' />
-        {countyCodes.map((county) => <option key={county.code} value={county.code}>{county.value}</option>)}
+        {counties.map((county) => <option key={county.code} value={county.code}>{county.value}</option>)}
       </SelectField>
     )
   }
 }
 
 CountySelectField.propTypes = {
-  countyCodes: PropTypes.array.isRequired,
+  counties: PropTypes.array.isRequired,
   gridClassName: PropTypes.string,
   id: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
