+++ spec/javascripts/reducers/systemCodes/countyCodesReducerSpec.js
@@ -1,9 +1,9 @@
 import {List, fromJS} from 'immutable'
 import * as matchers from 'jasmine-immutable-matchers'
-import countyCodesReducer from 'reducers/systemCodes/countyCodesReducer'
+import countiesReducer from 'reducers/systemCodes/countiesReducer'
 import {fetchSuccess} from 'actions/systemCodesActions'
 
-describe('countyCodesReducer', () => {
+describe('countiesReducer', () => {
   beforeEach(() => jasmine.addMatchers(matchers))
 
   describe('on FETCH_STATUS_CODES_SUCCESS', () => {
@@ -14,7 +14,7 @@ describe('countyCodesReducer', () => {
         {code: '789', value: 'Sacramento', category: 'address_county'},
         {code: 'ABC', value: 'San Diego', category: 'county_type'},
       ])
-      expect(countyCodesReducer(List(), action)).toEqualImmutable(fromJS([
+      expect(countiesReducer(List(), action)).toEqualImmutable(fromJS([
         {code: '123', value: 'San Francisco', category: 'county_type'},
         {code: '456', value: 'Monterey', category: 'county_type'},
         {code: 'ABC', value: 'San Diego', category: 'county_type'},
