+++ spec/javascripts/store/storeSpec.js
@@ -26,8 +26,8 @@ describe('Store', () => {
       contactStatuses: [],
       contactPurposes: [],
       communicationMethods: [],
-      countyCodes: [],
-      countyAgencies: {},
+      counties: [],
+      countyAgencies: [],
       locations: [],
       routing: {locationBeforeTransitions: null},
       investigationPeople: [],
